requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
cloudscraper==1.2.71
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
selenium==4.15.2
undetected-chromedriver==3.5.4
python-dotenv==1.0.0
pydantic==2.5.0
gspread==6.1.0
google-auth==2.23.4
